+++ app/src/main/java/com/unimind/unithing/Contract/SignUserContract.kt
@@ -15,4 +15,9 @@ interface SignUserContract {
         fun checkValidation(email: String)
     }
 
+    interface SignUserRepository {
+        fun requestSignUp(email: String, password: String,type: String, callback: (Boolean, String?) -> Unit)
+        fun requestSignIn(email: String, password: String, callback: (Boolean, String?) -> Unit)
+    }
+
 }
