+++ app/src/main/java/com/unimind/unithing/Repository/RemoteDataSource/CertificateRepositoryImpl.kt
@@ -1,18 +1,18 @@
-package com.unimind.unithing.Repository.UserRemoteDataSource
+package com.unimind.unithing.Repository.RemoteDataSource
 
 import android.graphics.Bitmap
-import android.util.Log
 import com.google.firebase.auth.ktx.auth
 import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.ktx.Firebase
 import com.google.firebase.storage.ktx.storage
+import com.unimind.unithing.Contract.CertificateContract
 import com.unimind.unithing.CustomApplication
-import com.unimind.unithing.Data.MajorCERT
+import com.unimind.unithing.Data.Certificate
 import com.unimind.unithing.R
 import com.unimind.unithing.StringResource
 import java.io.ByteArrayOutputStream
 
-object UserRepositoryImpl : UserRepository {
+object CertificateRepositoryImpl : CertificateContract.CertificateRepository {
 
     private val firebaseAuth = Firebase.auth
 
@@ -20,16 +20,17 @@ object UserRepositoryImpl : UserRepository {
         CustomApplication.ctx, R.string.db_major_certificate)
 
     private val firestoreCertDB = FirebaseFirestore.getInstance().collection(
-        certDBname)
+        certDBname
+    )
 
     private val userUid = firebaseAuth.uid.toString()
 
     private val firebaseStorage = Firebase.storage.reference
 
     override fun createCertificateDB(callback: (Boolean) -> Unit) {
-        val majorCERT = MajorCERT(null)
+        val certificate = Certificate(null)
 
-        firestoreCertDB.document(userUid).set(majorCERT)
+        firestoreCertDB.document(userUid).set(certificate)
             .addOnCompleteListener {
                 callback(true)
             }.addOnFailureListener {
@@ -38,7 +39,7 @@ object UserRepositoryImpl : UserRepository {
     }
 
     override fun uploadStorage(image: Bitmap, callback: (Boolean) -> Unit) {
-        val storageRef = firebaseStorage.child("${certDBname}/${userUid}.jpg")
+        val storageRef = firebaseStorage.child("$certDBname/$userUid.jpg")
 
         val baos = ByteArrayOutputStream()
         image.compress(Bitmap.CompressFormat.JPEG, 100, baos)
