--- app/src/main/java/com/unimind/unithing/Repository/SignUserRemoteDataSource/SignUserRepository.kt
@@ -1,6 +1,0 @@
-package com.unimind.unithing.Repository.SignUserRemoteDataSource
-
-interface SignUserRepository {
-    fun requestSignUp(email: String, password: String,type: String, callback: (Boolean, String?) -> Unit)
-    fun requestSignIn(email: String, password: String, callback: (Boolean, String?) -> Unit)
-}
